Added setTimeout around popup when connection closes (#17)
When a page is reloaded, the socket will be closed. If
you are editing and reloading HTML files, it could be
that you are frequently reloading the pages. In this
context, it is _not_ an error for the page to be reloaded,
so it is confusing if a red error message pops up on
the page.

This fix adds a `setTimeout` around the popup function
which will prevent the popup from closing when the page
is reloaded.

Closes #16